// src/app/core/api/model/product-photo.ts
/**
 * E-commerce platform API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.4.22
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Product } from './product';


export interface ProductPhoto { 
    product: Product;
    id: number;
    path: string;
    mimeType: string;
    thumbnailPath: string;
    placeholderBase64: string;
}